requirements.txt
streamlit==1.24.0
torch==2.2.2  # Latest compatible version
numpy==1.26.4
pandas==2.0.3
plotly==5.13.1
umap-learn==0.5.3
transformers==4.30.0
langchain==0.3.9
faiss-gpu==1.7.2  # or faiss-cpu==1.7.2 if you don't have a GPU
sentence-transformers==2.2.2
ragatouille==0.1.5
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0
tqdm==4.66.5
datasets==2.14.4